+++ tests/Equinox.Cosmos.Integration/JsonConverterTests.fs
@@ -46,6 +46,6 @@ type Base64ZipUtf8Tests() =
         let ser = JsonConvert.SerializeObject(e)
         test <@ ser.Contains("\"d\":\"") @>
         let des = JsonConvert.DeserializeObject<Store.Unfold>(ser)
-        let d = FsCodec.Core.IndexedEventData(-1L, false, des.c, des.d, null, DateTimeOffset.UtcNow)
+        let d = FsCodec.Core.TimelineEvent.Create(-1L, des.c, des.d, null)
         let decoded = unionEncoder.TryDecode d |> Option.get
         test <@ value = decoded @>